Exchange protocol works

+++ Homework_1/Merchant/src/main/java/merchant/repos/MerchantSessionsRepo.java
@@ -2,7 +2,6 @@ package merchant.repos;
 
 import merchant.entities.MerchantSessionsEntity;
 import org.springframework.data.repository.CrudRepository;
-import org.springframework.stereotype.Repository;
 
-public interface MerchantSessionsRepo extends CrudRepository<MerchantSessionsEntity, Long> {
+public interface MerchantSessionsRepo extends CrudRepository<MerchantSessionsEntity, Integer> {
 }
